requirements.txt
torch==2.2.2
numpy==1.26.3
matplotlib==3.7.1
pillow==10.2.0
torchvision==0.17.2
torchaudio
scipy==1.11.1
tensorboard==2.20.0     # used for logging
scikit-learn==1.7.1
open3d==0.19.0          # used for quaternion conversion
albumentations==2.0.8   # used for image augmentation
chardet==5.2.0          # fix requests warning
# kornia==0.8.1           # used for point cloud augmentation
